split out heartbeat and maintenance tests
