+++ grafana-plugin/e2e-tests/integrations/maintenanceMode/debugMode.test.ts
@@ -1,0 +1,25 @@
+import { test } from '../../fixtures';
+import { verifyThatAlertGroupIsRoutedCorrectlyButNotEscalated } from '../../utils/alertGroup';
+import { sendDemoAlert } from '../../utils/integrations';
+
+import {
+  createIntegrationAndEscalationChainAndEnableMaintenanceMode,
+  createRoutedText,
+  disableMaintenanceMode,
+} from './';
+
+test('debug mode', async ({ adminRolePage: { page, userName } }) => {
+  const { escalationChainName, integrationName } = await createIntegrationAndEscalationChainAndEnableMaintenanceMode(
+    page,
+    userName,
+    'Debug'
+  );
+  await sendDemoAlert(page);
+  await verifyThatAlertGroupIsRoutedCorrectlyButNotEscalated(
+    page,
+    integrationName,
+    createRoutedText(escalationChainName)
+  );
+
+  await disableMaintenanceMode(page, integrationName);
+});
